BE: 랭킹 동점 순위 처리 방법 수정

+++ backend/src/main/java/com/ssafy/newstock/memberstocks/service/MemberStocksService.java
@@ -105,7 +105,7 @@ public class MemberStocksService {
         DecimalFormat df = new DecimalFormat("0.0");
         String formattedROI = df.format(ROI);
         totalPrice=deposit+newPrice;
-        Long rank=redisService.getMemberRank(memberId);
+        Integer rank=redisService.getMemberRank(memberId);
         String rankSaveTime= redisService.getRankTime();
 
         return new AssetInfoResponse(nickname,totalPrice,deposit,profitAndLoss,formattedROI,rank,rankSaveTime);
